youtube-project/ranking-shorts-generator/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-socketio==5.10.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Video Processing
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.2.0
yt-dlp>=2024.10.7

# Scraping
TikTokApi==6.0.0
playwright==1.40.0
beautifulsoup4==4.12.2
httpx==0.26.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.9
